Refactor narrowphase, move free functions to separate files. Do not create contact constraints immediately after contact points are created; do it at the beginning of the step instead.
