+++ CMakeLists.txt
@@ -73,6 +73,7 @@ set(Edyn_SOURCES
     src/edyn/util/aabb_util.cpp
     src/edyn/util/moment_of_inertia.cpp
     src/edyn/util/shape_volume.cpp
+    src/edyn/util/collision_util.cpp
     src/edyn/shapes/triangle_mesh.cpp
     src/edyn/shapes/paged_triangle_mesh.cpp
     src/edyn/util/triangle_util.cpp
