+++ include/edyn/collision/narrowphase.hpp
@@ -8,24 +8,15 @@
 #include "edyn/comp/shape_index.hpp"
 #include "edyn/comp/position.hpp"
 #include "edyn/comp/orientation.hpp"
-#include "edyn/shapes/shapes.hpp"
 #include "edyn/collision/contact_manifold.hpp"
+#include "edyn/collision/contact_point.hpp"
 #include "edyn/collision/collision_result.hpp"
+#include "edyn/constraints/constraint_impulse.hpp"
+#include "edyn/util/collision_util.hpp"
 
 namespace edyn {
 
-struct contact_manifold;
 struct job;
-
-using body_view_t = entt::basic_view<entt::entity, entt::exclude_t<>, AABB, shape_index, position, orientation>; 
-using transform_view_t = entt::basic_view<entt::entity, entt::exclude_t<>, position, orientation>; 
-using contact_manifold_view_t = entt::basic_view<entt::entity, entt::exclude_t<>, contact_manifold>;
-
-void detect_collision(const contact_manifold &, collision_result &, 
-                      const body_view_t &, const tuple_of_shape_views_t &);
-void process_result(entt::registry &, entt::entity manifold_entity, 
-                    contact_manifold &, const collision_result &, 
-                    const transform_view_t &);
 
 class narrowphase {
     struct contact_point_construction_info {
@@ -38,6 +29,9 @@ class narrowphase {
         size_t count {0};
     };
 
+    void add_pending_contact(entt::entity contact_entity,
+                             std::array<entt::entity, 2> body);
+
 public:
     narrowphase(entt::registry &);
 
@@ -46,17 +40,30 @@ public:
     void update_async(job &completion_job);
     void finish_async_update();
 
+    /**
+     * @brief Detects and processes collisions for the given manifolds.
+     */
     template<typename Iterator>
     void update_contact_manifolds(Iterator begin, Iterator end);
 
-    template<typename Iterator>
-    void update_contact_manifolds(Iterator begin, Iterator end, 
-                                  contact_manifold_view_t &manifold_view);
+    template<typename ContactManifoldView, typename Iterator>
+    void update_contact_manifolds(Iterator begin, Iterator end,
+                                  ContactManifoldView &manifold_view);
+
+    /**
+     * @brief When new contact points are created, such as when
+     * `update_contact_manifolds` is called, contact constraints are not created
+     * immediately. This must be called to create contact constraints for new
+     * contact points. It gives the caller control over when the constraints are
+     * created.
+     */
+    void create_contact_constraints();
 
 private:
     entt::registry *m_registry;
     std::vector<contact_point_construction_info> m_cp_construction_infos;
     std::vector<contact_point_destruction_info> m_cp_destruction_infos;
+    std::vector<entt::entity> m_pending_contact_points;
 };
 
 template<typename Iterator>
@@ -65,19 +72,28 @@ void narrowphase::update_contact_manifolds(Iterator begin, Iterator end) {
     update_contact_manifolds(begin, end, manifold_view);
 }
 
-template<typename Iterator>
-void narrowphase::update_contact_manifolds(Iterator begin, Iterator end, 
-                                           contact_manifold_view_t &manifold_view) {
+template<typename ContactManifoldView, typename Iterator>
+void narrowphase::update_contact_manifolds(Iterator begin, Iterator end,
+                                           ContactManifoldView &manifold_view) {
     auto body_view = m_registry->view<AABB, shape_index, position, orientation>();
     auto tr_view = m_registry->view<position, orientation>();
+    auto cp_view = m_registry->view<contact_point>();
+    auto imp_view = m_registry->view<constraint_impulse>();
     auto views_tuple = get_tuple_of_shape_views(*m_registry);
 
     for (auto it = begin; it != end; ++it) {
-        entt::entity entity = *it;
-        auto &manifold = manifold_view.get(entity);
+        entt::entity manifold_entity = *it;
+        auto &manifold = manifold_view.get(manifold_entity);
         collision_result result;
-        detect_collision(manifold, result, body_view, views_tuple);
-        process_result(*m_registry, entity, manifold, result, tr_view);
+        detect_collision(manifold.body, result, body_view, views_tuple);
+
+        process_collision(manifold_entity, manifold, result, cp_view, imp_view, tr_view,
+                          [&] (const collision_result::collision_point &rp) {
+            auto contact_entity = create_contact_point(*m_registry, manifold_entity, manifold, rp);
+            add_pending_contact(contact_entity, manifold.body);
+        }, [&] (entt::entity contact_entity) {
+            destroy_contact_point(*m_registry, manifold_entity, contact_entity);
+        });
     }
 }
